CSP: fix compilation with MSVC

+++ epan/dissectors/packet-componentstatus.c
@@ -169,7 +169,7 @@ dissect_componentstatusprotocol_cspreport_message(tvbuff_t *message_tvb, proto_t
   proto_tree_add_item(message_tree, hf_cspreport_location,        message_tvb,  28, 128, ENC_UTF_8|ENC_NA);
   proto_tree_add_item(message_tree, hf_cspreport_status,          message_tvb, 156, 128, ENC_UTF_8|ENC_NA);
 
-  workload = 100.0 * CSR_GET_WORKLOAD(tvb_get_ntohs(message_tvb, 284));
+  workload = (float)(100.0 * CSR_GET_WORKLOAD(tvb_get_ntohs(message_tvb, 284)));
   if(workload < 0.0) {   /* Special value 0xffff -> -1.0 means "no load provided"! */
      proto_tree_add_float_format(message_tree, hf_cspreport_workload, message_tvb, 284, 2,
                                 workload, "Workload: N/A");
